Fix BEM class naming
Class names like `1/2`, `50` and `1-of-2` pollute the global namespace.
These classes describe cell widths, so make them be modifiers of the
`grid__cell` element:

- grid__cell--1/2
- grid__cell--50
- grid__cell--1-of-2

Modifiers should _modify_ a block or an element. They should not be used
by themselves. Consider the markup:

```
<div class="grid">
  <div class="1/2 1/4--wide grid__cell">Cell</div>
  <div class="1/2 1/4--wide grid__cell">Cell</div>
  <div class="1/2 1/4--wide grid__cell">Cell</div>
  <div class="1/2 1/4--wide grid__cell">Cell</div>
</div>
```
The `wide` modifier is being used on a `1/4` block, but the block itself
is not referenced. Instead it's understood that the `grid__cell` is
`1/2` until `wide` when it becomes `1/4`. The classes used in the above
example are disassociative.

Improve this concept:

```
<div class="grid">
  <div class="grid__cell grid__cell--1/2 grid__cell--1/4-for-wide">Cell</div>
  <div class="grid__cell grid__cell--1/2 grid__cell--1/4-for-wide">Cell</div>
  <div class="grid__cell grid__cell--1/2 grid__cell--1/4-for-wide">Cell</div>
  <div class="grid__cell grid__cell--1/2 grid__cell--1/4-for-wide">Cell</div>
</div>
```

This example more accurately expresses the width + responsive logic,
using BEM-like naming.